+++ src/egon/data/datasets/sanity_checks.py
@@ -8,7 +8,6 @@ Authors: @ALonso, @dana
 from math import isclose
 from pathlib import Path
 
-from loguru import logger
 from sqlalchemy import Numeric
 from sqlalchemy.sql import and_, cast, func, or_
 import matplotlib.pyplot as plt
@@ -16,8 +15,12 @@ import numpy as np
 import pandas as pd
 import seaborn as sns
 
-from egon.data import db
+from egon.data import db, logger
 from egon.data.datasets import Dataset
+from egon.data.datasets.electricity_demand_timeseries.cts_buildings import (
+    EgonCtsElectricityDemandBuildingShare,
+    EgonCtsHeatDemandBuildingShare,
+)
 from egon.data.datasets.emobility.motorized_individual_travel.db_classes import (
     EgonEvCountMunicipality,
     EgonEvCountMvGridDistrict,
@@ -51,7 +54,7 @@ class SanityChecks(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="SanityChecks",
-            version="0.0.4",
+            version="0.0.5",
             dependencies=dependencies,
             tasks={
                 sanitycheck_pv_rooftop_buildings,
@@ -60,6 +63,8 @@ class SanityChecks(Dataset):
                 etrago_eGon2035_heat,
                 residential_electricity_annual_sum,
                 residential_electricity_hh_refinement,
+                cts_electricity_demand_share,
+                cts_heat_demand_share,
             },
         )
 
@@ -82,8 +87,8 @@ def etrago_eGon2035_electricity():
     scn = "eGon2035"
 
     # Section to check generator capacities
-    print(f"Sanity checks for scenario {scn}")
-    print(
+    logger.info(f"Sanity checks for scenario {scn}")
+    logger.info(
         "For German electricity generators the following deviations between "
         "the inputs and outputs can be observed:"
     )
@@ -148,7 +153,7 @@ def etrago_eGon2035_electricity():
             sum_output.output_capacity_mw.sum() == 0
             and sum_input.input_capacity_mw.sum() == 0
         ):
-            print(
+            logger.info(
                 f"No capacity for carrier '{carrier}' needed to be"
                 f" distributed. Everything is fine"
             )
@@ -157,7 +162,7 @@ def etrago_eGon2035_electricity():
             sum_input.input_capacity_mw.sum() > 0
             and sum_output.output_capacity_mw.sum() == 0
         ):
-            print(
+            logger.info(
                 f"Error: Capacity for carrier '{carrier}' was not distributed "
                 f"at all!"
             )
@@ -166,7 +171,7 @@ def etrago_eGon2035_electricity():
             sum_output.output_capacity_mw.sum() > 0
             and sum_input.input_capacity_mw.sum() == 0
         ):
-            print(
+            logger.info(
                 f"Error: Eventhough no input capacity was provided for carrier"
                 f"'{carrier}' a capacity got distributed!"
             )
@@ -178,12 +183,12 @@ def etrago_eGon2035_electricity():
             ) * 100
             g = sum_input["error"].values[0]
 
-            print(f"{carrier}: " + str(round(g, 2)) + " %")
+            logger.info(f"{carrier}: " + str(round(g, 2)) + " %")
 
     # Section to check storage units
 
-    print(f"Sanity checks for scenario {scn}")
-    print(
+    logger.info(f"Sanity checks for scenario {scn}")
+    logger.info(
         "For German electrical storage units the following deviations between"
         "the inputs and outputs can be observed:"
     )
@@ -221,7 +226,7 @@ def etrago_eGon2035_electricity():
             sum_output.output_capacity_mw.sum() == 0
             and sum_input.input_capacity_mw.sum() == 0
         ):
-            print(
+            logger.info(
                 f"No capacity for carrier '{carrier}' needed to be "
                 f"distributed. Everything is fine"
             )
@@ -230,7 +235,7 @@ def etrago_eGon2035_electricity():
             sum_input.input_capacity_mw.sum() > 0
             and sum_output.output_capacity_mw.sum() == 0
         ):
-            print(
+            logger.info(
                 f"Error: Capacity for carrier '{carrier}' was not distributed"
                 f" at all!"
             )
@@ -239,7 +244,7 @@ def etrago_eGon2035_electricity():
             sum_output.output_capacity_mw.sum() > 0
             and sum_input.input_capacity_mw.sum() == 0
         ):
-            print(
+            logger.info(
                 f"Error: Eventhough no input capacity was provided for carrier"
                 f" '{carrier}' a capacity got distributed!"
             )
@@ -251,11 +256,11 @@ def etrago_eGon2035_electricity():
             ) * 100
             g = sum_input["error"].values[0]
 
-            print(f"{carrier}: " + str(round(g, 2)) + " %")
+            logger.info(f"{carrier}: " + str(round(g, 2)) + " %")
 
     # Section to check loads
 
-    print(
+    logger.info(
         "For German electricity loads the following deviations between the"
         " input and output can be observed:"
     )
@@ -305,7 +310,7 @@ def etrago_eGon2035_electricity():
 
     e = round((output_demand - input_demand) / input_demand, 2) * 100
 
-    print(f"electricity demand: {e} %")
+    logger.info(f"electricity demand: {e} %")
 
 
 def etrago_eGon2035_heat():
@@ -329,8 +334,8 @@ def etrago_eGon2035_heat():
     scn = "eGon2035"
 
     # Section to check generator capacities
-    print(f"Sanity checks for scenario {scn}")
-    print(
+    logger.info(f"Sanity checks for scenario {scn}")
+    logger.info(
         "For German heat demands the following deviations between the inputs"
         " and outputs can be observed:"
     )
@@ -370,11 +375,11 @@ def etrago_eGon2035_heat():
         * 100
     )
 
-    print(f"heat demand: {e_demand} %")
+    logger.info(f"heat demand: {e_demand} %")
 
     # Sanity checks for heat supply
 
-    print(
+    logger.info(
         "For German heat supplies the following deviations between the inputs "
         "and outputs can be observed:"
     )
@@ -404,7 +409,7 @@ def etrago_eGon2035_heat():
         round((heat_pump_output - heat_pump_input) / heat_pump_output, 2) * 100
     )
 
-    print(f"'central_heat_pump': {e_heat_pump} % ")
+    logger.info(f"'central_heat_pump': {e_heat_pump} % ")
 
     # Comparison for residential heat pumps
 
@@ -436,7 +441,7 @@ def etrago_eGon2035_heat():
         )
         * 100
     )
-    print(f"'residential heat pumps': {e_residential_heat_pump} %")
+    logger.info(f"'residential heat pumps': {e_residential_heat_pump} %")
 
     # Comparison for resistive heater
     resistive_heater_input = db.select_dataframe(
@@ -469,7 +474,7 @@ def etrago_eGon2035_heat():
         * 100
     )
 
-    print(f"'resistive heater': {e_resistive_heater} %")
+    logger.info(f"'resistive heater': {e_resistive_heater} %")
 
     # Comparison for solar thermal collectors
 
@@ -500,7 +505,7 @@ def etrago_eGon2035_heat():
         )
         * 100
     )
-    print(f"'solar thermal collector': {e_solar_thermal} %")
+    logger.info(f"'solar thermal collector': {e_solar_thermal} %")
 
     # Comparison for geothermal
 
@@ -529,7 +534,7 @@ def etrago_eGon2035_heat():
         round((output_geo_thermal - input_geo_thermal) / input_geo_thermal, 2)
         * 100
     )
-    print(f"'geothermal': {e_geo_thermal} %")
+    logger.info(f"'geothermal': {e_geo_thermal} %")
 
 
 def sanitycheck_pv_rooftop_buildings():
@@ -1155,7 +1160,8 @@ def sanitycheck_emobility_mit():
 
 
 def residential_electricity_annual_sum(rtol=1e-5):
-    """Sanity check for dataset electricity_demand_timeseries
+    """Sanity check for dataset electricity_demand_timeseries :
+    Demand_Building_Assignment
 
     Aggregate the annual demand of all census cells at NUTS3 to compare
     with initial scaling parameters from DemandRegio.
@@ -1188,14 +1194,15 @@ def residential_electricity_annual_sum(rtol=1e-5):
         verbose=False,
     )
 
-    print(
+    logger.info(
         "Aggregated annual residential electricity demand"
         " matches with DemandRegio at NUTS-3."
     )
 
 
 def residential_electricity_hh_refinement(rtol=1e-5):
-    """Sanity check for dataset electricity_demand_timeseries
+    """Sanity check for dataset electricity_demand_timeseries :
+    Household Demands
 
     Check sum of aggregated household types after refinement method
     was applied and compare it to the original census values."""
@@ -1229,4 +1236,58 @@ def residential_electricity_hh_refinement(rtol=1e-5):
         verbose=False,
     )
 
-    print("All Aggregated household types match at NUTS-3.")
+    logger.info("All Aggregated household types match at NUTS-3.")
+
+
+def cts_electricity_demand_share(rtol=1e-5):
+    """Sanity check for dataset electricity_demand_timeseries :
+    CtsBuildings
+
+    Check sum of aggregated cts electricity demand share which equals to one
+    for every substation as the substation profile is linearly disaggregated
+    to all buildings."""
+
+    with db.session_scope() as session:
+        cells_query = session.query(EgonCtsElectricityDemandBuildingShare)
+
+    df_demand_share = pd.read_sql(
+        cells_query.statement, cells_query.session.bind, index_col=None
+    )
+
+    np.testing.assert_allclose(
+        actual=df_demand_share.groupby(["bus_id", "scenario"])[
+            "profile_share"
+        ].sum(),
+        desired=1,
+        rtol=rtol,
+        verbose=False,
+    )
+
+    logger.info("The aggregated demand shares equal to one!.")
+
+
+def cts_heat_demand_share(rtol=1e-5):
+    """Sanity check for dataset electricity_demand_timeseries
+    : CtsBuildings
+
+    Check sum of aggregated cts heat demand share which equals to one
+    for every substation as the substation profile is linearly disaggregated
+    to all buildings."""
+
+    with db.session_scope() as session:
+        cells_query = session.query(EgonCtsHeatDemandBuildingShare)
+
+    df_demand_share = pd.read_sql(
+        cells_query.statement, cells_query.session.bind, index_col=None
+    )
+
+    np.testing.assert_allclose(
+        actual=df_demand_share.groupby(["bus_id", "scenario"])[
+            "profile_share"
+        ].sum(),
+        desired=1,
+        rtol=rtol,
+        verbose=False,
+    )
+
+    logger.info("The aggregated demand shares equal to one!.")