+++ CHANGELOG.rst
@@ -215,7 +215,10 @@ Added
 * Add distribution of CHP plants for eGon100RE
   `#851 <https://github.com/openego/eGon-data/issues/851>`_
 * Add sanity checks for residential electricity loads
-  `#915 <https://github.com/openego/eGon-data/pull/915>`_
+  `#902 <https://github.com/openego/eGon-data/issues/902>`_
+* Add sanity checks for cts loads
+  `#919 <https://github.com/openego/eGon-data/issues/919>`_
+``
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159
 .. _PR #703: https://github.com/openego/eGon-data/pull/703